backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
matplotlib
pytest
spacy
tika
#pyspark
fastapi
setuptools
PyMySQL
numpy
starlette
#ray
#ray[serve]
requests
pykwalify
PyYAML
faiss-cpu
sentence-transformers
elasticsearch
pandas
PyArrow
pydantic
typing
sqlmodel
sqlalchemy
uvicorn
rich
alembic
decorator
psycopg2
install-jdk
pypdf